refactor(spa-editor): share the attention predicate and the calendar-day formatter
Two surfaces now summarise the Connections section: the Settings shell's
banner and the section's own counters. The shell already learned once what
happens when each carries its own predicate — it shipped `error ||
needsReauth`, Wave 1 shipped `bridgeSourceStatus`, and the two disagreed on a
reachable case (a reachable probed source with no session: the predicate said
healthy, the card said amber). Git merged both without complaint.

`sourceNeedsAttention` moves to `application/connections/connection-source.ts`,
beside the type it tests, and `connection-attention.ts` re-exports it under the
name its own tests already use. `dayOf` becomes the shared `calendarDay` for
the same reason: both banners state the same date about the same source, and
two formatters would eventually disagree about which day that was.

No behaviour change and no test change — the exported names are unchanged,
which is the point of the re-export.
